Refactor to organise folders
