+++ gov_uk_dashboards/components/__init__.py
@@ -4,4 +4,4 @@ using different systems.
 Contains:
 - plotly: Module containing Plotly/Dash components
 """
-from . import plotly
+from . import dash
