+++ gov_uk_dashboards/components/dash/collapsible_panel.py
