+++ gov_uk_dashboards/components/dash/details.py
