+++ gov_uk_dashboards/components/helpers/generate_dash_graph_from_figure.py
