+++ gov_uk_dashboards/components/plotly/choropleth_map.py
@@ -4,7 +4,7 @@ import polars as pl
 import plotly.graph_objects as go
 from dash import dcc
 from gov_uk_dashboards import colours
-from gov_uk_dashboards.components.display_chart_or_table_with_header import (
+from gov_uk_dashboards.components.helpers.display_chart_or_table_with_header import (
     display_chart_or_table_with_header,
 )
